% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair dois números}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{Um número}

\item{b}{Outro número}
}
\description{
Esta função serve para subtrair dois números a e b
}
\examples{
subtrari(6,2)

}
